+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.4.17
 - Add support for ntfy
+- Add Rescan of changed media
 - Fix handling of corrupt video
   
 ### v1.4.16